+++ resources/web/wwi/protoVisualizer/FieldModel.js
@@ -1,4 +1,4 @@
-import {VRML} from './constants.js';
+import {VRML} from './Vrml.js';
 
 export const FieldModel = {
   'Gyro': {
