+++ resources/web/wwi/protoVisualizer/Vrml.js
@@ -1,6 +1,5 @@
 'use strict';
 
-import {VRML} from './constants.js';
 import {Node} from './Node.js';
 
 class SingleValue {
@@ -850,3 +849,24 @@ export function stringifyType(type) {
       throw new Error('Unknown VRML type: ', type);
   }
 }
+
+export const VRML = {
+  MFBool: 0,
+  SFBool: 1,
+  MFColor: 2,
+  SFColor: 3,
+  MFFloat: 4,
+  SFFloat: 5,
+  MFInt32: 6,
+  SFInt32: 7,
+  MFNode: 8,
+  SFNode: 9,
+  MFRotation: 10,
+  SFRotation: 11,
+  MFString: 12,
+  SFString: 13,
+  MFVec2f: 14,
+  SFVec2f: 15,
+  MFVec3f: 16,
+  SFVec3f: 17
+};
