--- resources/web/wwi/protoVisualizer/constants.js
@@ -1,22 +1,0 @@
-'use strict';
-
-export const VRML = {
-  MFBool: 0,
-  SFBool: 1,
-  MFColor: 2,
-  SFColor: 3,
-  MFFloat: 4,
-  SFFloat: 5,
-  MFInt32: 6,
-  SFInt32: 7,
-  MFNode: 8,
-  SFNode: 9,
-  MFRotation: 10,
-  SFRotation: 11,
-  MFString: 12,
-  SFString: 13,
-  MFVec2f: 14,
-  SFVec2f: 15,
-  MFVec3f: 16,
-  SFVec3f: 17
-};
