--- resources/web/wwi/protos/RegeneratingShape.proto
@@ -1,25 +1,0 @@
-#VRML_SIM R2023a utf8
-# template language: javascript
-
-PROTO RegeneratingShape [
-  field SFBool flag1 FALSE
-]
-{
-  Shape {
-    %< if (fields.flag1.value) { >%
-      appearance PBRAppearance {
-        baseColor 1 0 0
-      }
-      geometry Sphere {
-        radius 0.5
-      }
-    %< } else { >%
-      appearance PBRAppearance {
-        baseColor 0 0 1
-      }
-      geometry Box {
-        size 1 1 1
-      }
-    %< } >%
-  }
-}
