+++ resources/web/wwi/test.html
@@ -1,18 +1,15 @@
 <!DOCTYPE html>
 
 <style>
-  html,
-  body {
+  html, body {
     margin: 0;
     height: 100%;
     width: 100%;
   }
 </style>
-
 <head>
 
 </head>
-
 <body>
   <div style="height:100%; width:100%" id=webots-container></div>
 </body>
